==> 01_advanced_pandas_preparing_data.ipynb <==
# Advanced Pandas - Preparing Data
**Abid Ali**

Skype: Abd.Soft

Email: [abdsoftfsd@gmail.com](mailto:abdsoftfsd@gmail.com)

"The information used herein was obtained free of charge from and is
copyrighted by the Hockey Databank project.  For more information about the
Hockey Databank project please visit
[http://sports.groups.yahoo.com/group/hockey-databank](http://sports.groups.yahoo.com/group/hockey-databank)"

import os            # for file paths
import pandas as pd
from pandas import isnull


# Master.csv
master = pd.read_csv(os.path.join('data','Master.csv'))
master.head()


master.shape


master.columns


# Braces let us structure the code in a more readable way
(master['playerID']
    .pipe(isnull)
    .value_counts())


# This is an alternative to:
isnull(master['playerID']).value_counts()


master_original = master.copy()


master.dropna(subset=['playerID'], inplace=True)
master.shape


master.columns


# Drop rows where no data for NHL years
master = master.dropna(subset=['firstNHL', 'lastNHL'], how='all')
master.shape


master.dtypes



master = master.loc[master['lastNHL'] >= 1980]
master.shape

master.columns


columns_to_keep = ['playerID', 'firstName', 'lastName', 'pos',
                   'birthYear', 'birthMon', 'birthDay', 'birthCountry', 'birthState',
                   'birthCity']
master[columns_to_keep].head()


master.filter(columns_to_keep).head()


master.shape


master_modified = master.copy()
master = master.filter(columns_to_keep)
master.shape


master = master_modified.copy()
master.shape


master.filter(regex="(playerID|pos|^birth)|(Name$)").columns

master.filter(regex="(playerID|pos|^birth|Name$)").columns


master = master.filter(regex="(playerID|pos|^birth|Name$)")
master.shape


master.columns


def mem_mib(df):
    print("{0:.2f} MiB".format(
        df.memory_usage().sum() / (1024 * 1024)
    ))

mem_mib(master)
mem_mib(master_original)


master["pos"].value_counts()


pd.Categorical(master["pos"])


def make_categorical(df, col_name):
    df.loc[:, col_name] = pd.Categorical(df[col_name])


make_categorical(master, "pos")
make_categorical(master, "birthCountry")
make_categorical(master, "birthState")
mem_mib(master)


master.head()


master = master.set_index('playerID')
master.head()


master.to_pickle(os.path.join('data','modified', 'master.pickle'))
master.to_csv(os.path.join('data', 'modified', 'master.csv'))


# Scoring.csv


original_scoring = pd.read_csv(os.path.join('data', 'Scoring.csv'))
scoring = original_scoring.copy()
mem_mib(scoring)
scoring.shape


def recent_nhl_only(df):
    return df[(df["lgID"] == "NHL") & (df["year"] >= 1980)]

scoring = recent_nhl_only(scoring)
scoring.shape


scoring.columns


scoring = scoring.filter(regex="^(?!(Post|PP|SH)).*")
scoring.head()

cols = scoring.columns
cols

cols[[0,1,3,6,7,8,9,14]]


cols_to_keep = ['playerID', 'year', 'tmID', 'GP', 'G', 'A', 'Pts', 'SOG']
scoring = scoring.filter(cols_to_keep)
scoring.head()


scoring['tmID'].value_counts()

make_categorical(scoring, 'tmID')
scoring.head()


scoring.reset_index().head()


scoring.reset_index(drop=True, inplace=True)
scoring.head()

scoring.to_pickle(os.path.join('data','modified','scoring.pickle'))
scoring.to_csv(os.path.join('data','modified','scoring.csv'))


# Team.csv


teams = pd.read_csv(os.path.join('data', 'Teams.csv'))
teams.shape


teams.columns


teams = recent_nhl_only(teams)
teams = teams[['year','tmID', 'name']]
teams.head()


teams.shape


teams.nunique()


make_categorical(teams, 'tmID')


teams.to_pickle(os.path.join('data', 'modified', 'teams.pickle'))
teams.to_csv(os.path.join('data', 'modified', 'teams.csv'))


# TeamSplits.csv


team_splits = pd.read_csv(os.path.join('data', 'TeamSplits.csv'))
team_splits.shape


team_splits.columns


team_splits = recent_nhl_only(team_splits)


team_splits.columns


cols_to_drop = team_splits.columns[3:11]
cols_to_drop


team_splits = team_splits.drop(columns=cols_to_drop)


team_splits.columns


# rows can be dropped by row_labels
# df.drop(rows=row_labels)    # <- to drop rows

team_splits = team_splits.drop(columns="lgID")
team_splits.columns


make_categorical(team_splits, 'tmID')


team_splits.to_pickle(os.path.join('data', 'modified', 'team_splits.pickle'))
team_splits.to_csv(os.path.join('data', 'modified', 'team_splits.csv'))
