==> 02_joining_multiple_data_tables.ipynb <==
# Advanced Pandas - Joining Multiple Data Tables
**Abid Ali**

Skype: Abd.Soft

Email: [abdsoftfsd@gmail.com](mailto:abdsoftfsd@gmail.com)


"The information used herein was obtained free of charge from and is
copyrighted by the Hockey Databank project.  For more information about the
Hockey Databank project please visit
[http://sports.groups.yahoo.com/group/hockey-databank](http://sports.groups.yahoo.com/group/hockey-databank)"

import pandas as pd
import os


master = pd.read_pickle(os.path.join('data', 'modified', 'master.pickle'))
scoring = pd.read_pickle(os.path.join('data', 'modified', 'scoring.pickle'))
teams = pd.read_pickle(os.path.join('data', 'modified', 'teams.pickle'))
team_splits = pd.read_pickle(os.path.join('data', 'modified', 'team_splits.pickle'))


master.head(2)


scoring.head(2)


teams.head(2)


team_splits.head(2)


pd.merge(left=master, right=scoring, left_index=True, right_on='playerID').head()


# where is the resulting dataframe indexing coming from?
# - Answer: from the right table.
scoring.index


scoring.index + 3


scoring.index = scoring.index + 3


scoring.head()


pd.merge(master, scoring, left_index=True, right_on="playerID").head()

index starts with 3, so it is coming from the right table, as we have modified the right table index to start from 3.
Left index is just dropped.
We have playerID column, but it came from the right table.

# merging index on index
pd.merge(master, scoring.set_index('playerID', drop=True), left_index=True, right_index=True).head()


scoring = scoring.reset_index(drop=True)
pd.merge(master, scoring, left_index=True, right_on="playerID").head()


print(
    pd.merge(master, scoring, left_index=True, right_on="playerID").shape,
    pd.merge(master, scoring, left_index=True, right_on="playerID", how="right").shape
)


master = pd.read_pickle(os.path.join('data', 'modified', 'master.pickle'))
print(master.shape,scoring.shape)

# Drop random records
master2 = master.drop(master.sample(5).index)

print(
    'Inner Join: ',
    pd.merge(master2, scoring, left_index=True, right_on='playerID').shape,
    '\nRight Join: ',
    pd.merge(master2, scoring, left_index=True, right_on='playerID', how="right").shape
)


merged = pd.merge(master2, scoring, left_index=True, right_on="playerID", how="right", indicator=True)
merged.head()


merged.shape


merged['_merge'].value_counts()


merged[merged["_merge"] == "right_only"].head()

scoring2 = scoring.drop(scoring.sample(100).index)
merged = pd.merge(master2, scoring2, left_index=True, right_on="playerID", how="outer", indicator=True)


merged[(merged['_merge'] == 'left_only') | (merged['_merge'] == 'right_only')].sample(10)


merged[merged['_merge'].str.endswith('only')].sample(10)


pd.merge(master, scoring, left_index=True, right_on="playerID", validate="1:m").head()


try:
    pd.merge(master, scoring, left_index=True, right_on='playerID',
             validate="1:1")
except Exception as e:
    print(e)


merged = pd.merge(master, scoring, left_index=True, right_on="playerID")
merged.head()


merged.columns


merged = merged.filter(regex="^(?!(birth)).*")
merged.head()

merged.to_pickle(os.path.join("data", 'modified', 'scoring_merged.pickle'))
merged.to_csv(os.path.join("data", 'modified', 'scoring_merged.csv'))


teams.sample(2)


team_splits.sample(2)


teams2 = teams[['tmID', 'name']]
teams2 = teams2.drop_duplicates()


teams2['tmID'].value_counts().head()


teams2.loc[teams2['tmID'] == "CHI"]


teams2 = teams2[teams2['tmID'] == 'CHI']
team_splits2 = team_splits[team_splits['tmID'] == 'CHI'].sample(2)
team_splits2


pd.merge(teams2, team_splits2, on='tmID')


pd.merge(team_splits, teams, left_on=['tmID', 'year'], right_on=['tmID', 'year']).head(4)


try:
    result = pd.merge(team_splits, teams,
             left_on=['tmID', 'year'],
             right_on=['year', 'tmID']).head(4)
    print(result)
except Exception as e:
    print(e)

team_splits.dtypes


team_splits.dtypes


pd.merge(team_splits, teams, on=["tmID", 'year']).head(4)


# pd.merge(team_splits, teams, on=["tmID", 'year'])[['tmID', 'year']].value_counts()
pd.merge(team_splits, teams, on=["tmID", 'year']).head(5)

pd.merge(team_splits, teams, on=['tmID'], suffixes=('_team_splits', '_teams')).head(4).filter(like="year")


merged = pd.merge(team_splits, teams, on=['tmID', 'year'])
merged.head()


merged.to_pickle(os.path.join('data', 'modified', 'team_splits_merged.pickle'))
merged.to_csv(os.path.join('data', 'modified', 'team_splits_merged.csv'))
