==> 04_multi_index.ipynb <==
# Advanced Pandas - Multi Index
**Abid Ali**

Skype: Abd.Soft

Email: [abdsoftfsd@gmail.com](mailto:abdsoftfsd@gmail.com)

"The information used herein was obtained free of charge from and is
copyrighted by the Hockey Databank project.  For more information about the
Hockey Databank project please visit
[http://sports.groups.yahoo.com/group/hockey-databank](http://sports.groups.yahoo.com/group/hockey-databank)"

import pandas as pd
import os


master = pd.read_pickle(os.path.join('data','modified', 'master.pickle'))
scoring = pd.read_pickle(os.path.join('data','modified', 'scoring_merged.pickle'))
team_splits = pd.read_pickle(os.path.join('data','modified', 'team_splits_merged.pickle'))


scoring.head()


mi = scoring.set_index(['playerID', 'year'])
mi.head()


mi.index


len(mi.index.levels)


mi.index.nlevels


mi.index.levels[0]


mi.index.levels[1]


mi.groupby(level='year')['G'].max().head()


mi.groupby(level='year')['G'].idxmax().head()


mi.loc[mi.groupby(level='year')['G'].idxmax()][['firstName', 'lastName', 'G', 'A']].head(20)


mi.loc[('gretzwa01', 1982), :]


idx = pd.IndexSlice
mi2 = mi.sort_index()


for slicing to work correctly, index must be sorted lexicographically first.


# mi.index.is_lexsorted()
# mi.index.is_monotonic_decreasing
mi.index.is_monotonic_increasing

sliced = mi.loc[idx[:, 1997:2000], :]
sliced.head()


# range from 1997 to 2004, with a step of 2. (ending range is inclusive in pandas)
sliced = mi.loc[idx['aaltoan01': 'adamscr02', 1997:2004:2], :]
sliced.head()


# what if we want to select records from 1997 to 2000, and then 2004 to 2006
# no such built-in function available, so we can write our own



locs = mi.index.get_locs(
    (idx['aaltoan01': 'adamscr01', 1997:2000])
)
locs


sliced = mi.iloc[locs, :]
sliced.head()


import numpy as np

def get_many_locs(df, slices):
    arr = np.empty(0, dtype='int')
    for s in slices:
        locs = df.index.get_locs((s))
        arr = np.concatenate((arr, locs))
    return arr

locs = get_many_locs(
    mi,
    (
        idx['aaltoan01': 'adamscr01', 1997:2000],
        idx['aaltoan01': 'adamscr01', 2004:2006]
    )
)
locs


sliced = mi.iloc[locs, :]
sliced.head(10)


mi = mi.sort_index(level='year')
mi.head()


mi.index.is_monotonic_increasing


try:
    mi.loc[idx[:, 1997:2000], :]
except Exception as e:
    print(e)


swapped = mi.swaplevel()
swapped.head()


swapped.index.is_monotonic_increasing


# Wide and long formats revisited


team_splits.head(5)


team_splits = team_splits.drop(columns=['tmID'])
team_splits = team_splits.set_index(['year', 'name'])
team_splits.head()


months = team_splits.columns.map(lambda x: x[0:3])
metrics = team_splits.columns.map(lambda x: x[3:len(x)])

print(months)
print(metrics)


months = team_splits.columns.map(lambda x: x[0:3])
metrics = team_splits.columns.map(lambda x: x[3:])

print(months)
print(metrics)

midx = pd.MultiIndex.from_arrays([months, metrics])
midx


team_splits.columns = midx
team_splits.head()

team_splits = team_splits.stack(level=0)
team_splits.head(10)


team_splits = team_splits.swaplevel(1, 2)
team_splits.head()


team_splits = team_splits.reorder_levels([2,0,1])
team_splits.head()


team_splits.index.set_names('month', level=2, inplace=True)
team_splits.head()


team_splits.groupby(level=['name', 'year']).sum().head(10)


idx = pd.IndexSlice

team_splits.index.is_monotonic_increasing

team_splits.sort_index(inplace=True)
team_splits.index.is_monotonic_increasing

team_splits.loc[idx['Anaheim Ducks', 2006]]


team_splits.index


unstacked = team_splits.unstack(level=['year', 'month'])


unstacked.head()


unstacked.index


unstacked.index.nlevels


unstacked.columns


unstacked.columns.names


unstacked.columns.names = ['metrics', 'year', 'month']


unstacked.head()


unstacked_copy = unstacked.copy()


unstacked.columns = unstacked.columns.reorder_levels([1,2,0])
unstacked.head()


team_splits.to_pickle(os.path.join('data', 'modified', 'team_splits_mi.pickle'))
mi.to_pickle(os.path.join('data', 'modified', 'scoring_mi.pickle'))

team_splits.to_csv(os.path.join('data', 'modified', 'team_splits_mi.csv'))
mi.to_csv(os.path.join('data', 'modified', 'scoring_mi.csv'))


unstacked.to_csv(os.path.join('data', 'modified', 'team_splits_unstacked.csv'))
