==> 05_handling_time_series.ipynb <==
# Advanced Pandas - Handling Time Series
**Abid Ali**

Skype: Abd.Soft

Email: [abdsoftfsd@gmail.com](mailto:abdsoftfsd@gmail.com)

"The information used herein was obtained free of charge from and is
copyrighted by the Hockey Databank project.  For more information about the
Hockey Databank project please visit
[http://sports.groups.yahoo.com/group/hockey-databank](http://sports.groups.yahoo.com/group/hockey-databank)"

import pandas as pd
import os

master = pd.read_pickle(os.path.join('data','modified', 'master.pickle'))

master.head()


birthInfo = master.loc[:, ['birthYear', 'birthMon', 'birthDay']]
birthInfo.head()

(birthInfo.loc[birthInfo['birthYear'] == 1970]
    .sample(3)
    .groupby(['birthYear', 'birthMon'])
    .count()
 )

master = pd.read_pickle(os.path.join('data', 'modified', 'master.pickle'))
team_splits = pd.read_pickle(os.path.join('data', 'modified', 'team_splits_mi.pickle'))


team_splits.head()


master.head()


master = master.assign(birthDate = pd.to_datetime({
    'year': master.birthYear,
    'month': master.birthMon,
    'day': master.birthDay
}))
master = master.drop(columns=['birthYear', 'birthMon', 'birthDay'])
master.head()


master.birthDate.head(2)


master.birthDate[0]


ts = master.birthDate[0]


print(u"{}, {} {} {}".format(ts.day_name(), ts.day, ts.month_name(), ts.year))


tsz = ts.tz_localize('America/Toronto')
tsz


tsz.tz_convert('Europe/London')


years = master.birthDate.dt.year
years.head()

strings = master.birthDate.dt.strftime("%a-%d-%m-%Y")
strings.head()


strings = master.birthDate.dt.strftime('%Y-%m-%d')
strings.head()


dates = pd.to_datetime(strings)
dates.head()


date = dates[0]
date.month_name()


date.year

date.month


date.day_name()


strings = strings.str.replace('-', 'xx')
strings.head()


try:
    pd.to_datetime(strings)
except Exception as e:
    print(e)


parsed_dates = pd.to_datetime(strings, format="%Yxx%mxx%d")
parsed_dates.head()


strings_error = strings.copy()
strings_error.iloc[34] = "xx123"

try:
    pd.to_datetime(strings_error, format="%Yxx%mxx%d")
except Exception as e:
    print(e)


parsed = pd.to_datetime(strings_error, format="%Yxx%mxx%d", errors="coerce")
parsed.iloc[34]


parsed.head()


parsed = pd.to_datetime(strings_error, format="%Yxx%mxx%d", errors="ignore")
parsed.head()

birth_dates = pd.Series(data=master.index, index=master.birthDate)
birth_dates.head(2)


birth_dates.index


birth_dates = birth_dates.sort_index()
birth_dates.head(3)


born_1980 = birth_dates['1980']
born_1980.head()


birth_dates['1980-7']

birth_dates['1980-7-24']


birth_dates['1980-7-20': '1980-7-31']


diff = birth_dates.index[1] - birth_dates.index[0]
diff


birth_dates.index[10]


birth_dates.index[10] + diff


birth_dates.index - birth_dates.index[0]


diff.components


diff = birth_dates.index[1] - birth_dates.index[0]
diff = diff + '5 hours' - '10 minutes'
diff

diff.round('10 D')


diff.round('1H')


birth_dates.head()
# birth_dates = birth_dates.to_period(freq="D")

birth_dates = birth_dates.to_period(freq="D")

birth_dates.head()

birth_dates.index


birth_dates.index[0]


start = birth_dates.index[0].start_time
start

end = birth_dates.index[0].end_time
end

start < (start + pd.Timedelta('5H')) < end


start < (start + pd.Timedelta('23H59T59S')) < end


start < (start + pd.Timedelta('23H59T60S')) < end

birth_dates.index.year


birth_dates['1975'].head(2)


# Advanced Strategies


birth_dates.resample('1M')


birth_dates.resample('1M').count().head()


birth_dates.resample('2M').count().head()


birth_dates.resample('1Q').count().head(2)


birth_dates.resample('W').count().head()


from pandas.tseries.offsets import BDay
p = birth_dates.index[2]
print(p.to_timestamp())
print(p.to_timestamp() + BDay(7))


# shifting index by 1 day
birth_dates.shift(1, freq="D").head()


# if we don't add freq, it will shift values instead of index.
birth_dates.shift(1).head()


team_splits = team_splits.sort_index(level='year')
team_splits.head()

team_splits = team_splits.reset_index()
team_splits.head()


string_date = team_splits.apply(lambda x:
                                pd.Period(f"{x.year}-{x.month}"), axis=1)
string_date[0]


team_splits = team_splits.assign(month = string_date)
team_splits = team_splits.drop('year', axis=1)
team_splits.head()


team_splits = team_splits.set_index(['month'])


team_splits.head()


team_splits.index


quarter_sums = team_splits.groupby('name').resample("Q").sum()
quarter_sums.head()


quarter_means = team_splits.groupby('name').resample("Q").mean()
quarter_means.head()

team_splits.to_pickle(os.path.join('data', 'modified', 'team_splits_periods.pickle'))
birth_dates.to_pickle(os.path.join('data', 'modified', 'birth_dates.pickle'))

team_splits.to_csv(os.path.join('data', 'modified', 'team_splits_periods.csv'))
birth_dates.to_csv(os.path.join('data', 'modified', 'birth_dates.csv'))
